Changed double values to decimal type values.

+++ TraderDashboardUi.Entity/Position.cs
@@ -15,17 +15,17 @@ namespace TraderDashboardUi.Entity
 
         public int Units { get; set; }
 
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         public string BidPrice { get; set; }
         public string AskPrice { get; set; }
 
-        public double BackTestPipStopLoss { get; set; }
-        public double BackTestPipTakeProfit { get; set; }
+        public decimal BackTestPipStopLoss { get; set; }
+        public decimal BackTestPipTakeProfit { get; set; }
         public string BackTestBuySell { get; set; }
 
-        public double BackTestClosePositionPrice { get; set; }
-        public double BackTestClosePositionPL { get; set; }
+        public decimal BackTestClosePositionPrice { get; set; }
+        public decimal BackTestClosePositionPL { get; set; }
 
         public bool BackTestActive { get; set; } = false;
